fix(charts): Lookup charts by chart name, not release
The status quo is to look up charts by their release name, which *may*
be the same as the chart name, but often differs.

This changes the lookup to use the internal chart name, which will be
what ArtifactHub or Helm index.yaml would use.

Also, added the chart name and namespace to the web ui, as well as in
the pretty printed table mode, and for the metrics.